add fileManage
