+++ WebContent/share.jsp
@@ -11,7 +11,7 @@
     <link href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css" media="all" rel="stylesheet"
         type="text/css" />
 
-    <script src="https://code.jquery.com/jquery-3.2.1.min.js"></script>
+    <script src="js/jquery-3.2.1.min.js"></script>
 
     <script src="js/fileinput.min.js" type="text/javascript"></script>
     <script src="themes/fa/theme.js" type="text/javascript"></script>
