--- src/com/clouddrive/biz/DelFileBiz.java
@@ -1,5 +1,0 @@
-package com.clouddrive.biz;
-
-public interface DelFileBiz {
-	public boolean delFile(String path, String uuidName);
-}
