+++ src/com/clouddrive/biz/FileManageBiz.java
@@ -1,0 +1,9 @@
+package com.clouddrive.biz;
+
+public interface FileManageBiz {
+	// 删除文件
+	public boolean delFile(String path, String uuidName);
+	// 新建文件夹
+	public boolean mkdirFolder(String savePath, String fileName, String userName);
+	
+}
