+++ src/com/clouddrive/servlet/DownLoadServlet.java
@@ -10,11 +10,10 @@ import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import javax.servlet.http.HttpSession;
 
 import com.clouddrive.dao.impl.ShareDaoImpl;
-import com.clouddrive.entity.FileMessage;
 
+@SuppressWarnings("serial")
 public class DownLoadServlet extends HttpServlet {
 	@Override
 	protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
