+++ src/com/clouddrive/servlet/LoginServlet.java
@@ -11,6 +11,7 @@ import javax.servlet.http.HttpSession;
 import com.clouddrive.biz.impl.LoginBizImpl;
 import com.clouddrive.entity.User;
 
+@SuppressWarnings("serial")
 public class LoginServlet extends HttpServlet {
 	@Override
 	protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
