+++ src/com/clouddrive/servlet/LogoutServlet.java
@@ -8,6 +8,7 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 
+@SuppressWarnings("serial")
 public class LogoutServlet extends HttpServlet {
 	@Override
 	protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
@@ -15,7 +16,6 @@ public class LogoutServlet extends HttpServlet {
 		session.removeAttribute("name");
 		session.removeAttribute("path");
 		resp.sendRedirect("/CloudDrive/home.jsp");
-		return ;
 	}
 	
 	@Override
