+++ src/com/clouddrive/servlet/RegisterServlet.java
@@ -1,6 +1,7 @@
 package com.clouddrive.servlet;
 
 import java.io.IOException;
+
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
@@ -8,9 +9,7 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 
-import com.clouddrive.biz.impl.LoginBizImpl;
 import com.clouddrive.biz.impl.RegisterBizImpl;
-import com.clouddrive.entity.User;
 
 /**
  * Servlet implementation class RegisterServlet
